Cleanup

+++ acc_simulator/cli/acc_sim.py
@@ -121,7 +121,7 @@ def llama_eval(
 
     transformers.set_seed(0)
     # TODO: set the model path to the models checkpoints already stored inside .data/models/hw1020/
-    tokenizer, model = setup_model(model_name, model_parallel, dtype=torch.float16, 
+    tokenizer, model = setup_model(model_name, model_parallel, dtype=torch.bfloat16, 
                                    device=device_id if not model_parallel else None)
     # breakpoint()
     model.eval()
